Accumulate measurement value in double, seems marginally faster
Benchmark                                                           Mode  Cnt    Score    Error   Units
DoubleParsingBenchmark.ourToDouble                                 thrpt   10  569.771 ±  7.065  ops/us
DoubleParsingBenchmark.ourToDoubleAccumulateInToDouble             thrpt   10  648.026 ±  7.741  ops/us
DoubleParsingBenchmark.ourToDoubleDivideInsteadOfMultiply          thrpt   10  570.412 ±  9.329  ops/us
DoubleParsingBenchmark.ourToDoubleNegative                         thrpt   10  512.618 ±  8.580  ops/us
DoubleParsingBenchmark.ourToDoubleNegativeAccumulateInToDouble     thrpt   10  565.043 ± 18.137  ops/us
DoubleParsingBenchmark.ourToDoubleNegativeDivideInsteadOfMultiply  thrpt   10  511.228 ± 13.967  ops/us
DoubleParsingBenchmark.stringToDouble                              thrpt   10   52.310 ±  1.351  ops/us
DoubleParsingBenchmark.stringToDoubleNegative                      thrpt   10   50.785 ±  1.252  ops/us

+++ src/main/java/dev/morling/onebrc/CalculateAverage_palmr.java
@@ -143,7 +143,7 @@ public class CalculateAverage_palmr {
             }
             else if (currentChar == '\n') {
                 if (state.stationPointerEnd != 0) {
-                    double value = (double) state.measurementValue * state.exponent;
+                    double value = state.measurementValue * state.exponent;
 
                     MeasurementAggregator aggregator = state.aggregators.computeIfAbsent(state.stationBuffer, state.stationPointerEnd, state.signedHashCode);
                     aggregator.count++;
@@ -182,7 +182,7 @@ public class CalculateAverage_palmr {
         byte[] stationBuffer = new byte[STATION_NAME_BUFFER_SIZE];
         int signedHashCode = 0;
         int stationPointerEnd = 0;
-        long measurementValue = 0;
+        double measurementValue = 0;
         double exponent = 0.1;
 
         public void reset() {
